debian: Include SPIR-V (*.spv) outputs in libclc package
This adds the following to the libclc-12 package:

/usr/lib/clc/spirv64-mesa3d-.spv
/usr/lib/clc/spirv-mesa3d-.spv

Since Debian buster & Ubuntu 18.04 don't have the llvm-spirv package,
they cannot support this feature. Therefore, as suggested by tjaalton,
we make the control Build-Depends for spirv dependencies include
"| hello".

This causes the simple "hello world" package to be installed on these
older distros, which is small and unrelated to llvm. Then, in
debian/rules, we check to see if llvm-spirv is installed.

Note that sbuild must use `--resolve-alternatives` for these older
distributions, or sbuild won't search for the "hello" alternatives in
the Build-Depends.

Signed-off-by: Jordan Justen <jljusten@debian.org>