Update wasm-sysroot-usr.diff to fix #1029010
Update the wasm-sysroot-usr.diff patch to restore functionality that was
accidentally dropped when the patch was forward-ported from 14 to 15.
This resolves an issue in which clang++ builds would fail if
libc++-15-dev was installed alongside libc++-15-dev-wasm32.

This issue was rightfully caught in the autopkgtests, and has prevented
the package from migrating to testing, something that should hopefully
be resolved now.

Closes: #1029010, #1032317