Convert libclang-common-*-dev to Arch: all/M-A: foreign
libclang-common-*-dev is an Architecture: any package, possibly due to
the fact that it contained compiler-rt before it was split off to a
separate package. It is not marked as Multi-Arch: same, despite the
hinter suggesting to do so, presumably due to identical files.

Convert it to Architecture: all / Multi-Arch: foreign instead. This
requires the removal of the arch: any dependency libllvm17, which is
spurious anyway. However, for extra safety, add version specifiers to
other packages that do depend on (an unversioned) libllvm17, such as
clang.