Fix i386 FTBFS resulting from COMPILER_RT_DEFAULT_TARGET_ONLY
> dh_install: warning: Cannot find (any matches for) "usr/lib/llvm-17/lib/clang/17/bin/hwasan_symbolize" (tried in ., debian/tmp)

hwasan_symbolize is actually an 64-bit-only feature, and was
accidentally included before. Replace the arch opt-out list, with an
opt-in one and a comment pointing to HWASAN_SUPPORTED_ARCH.

> dh_install: warning: Cannot find (any matches for) "usr/lib/llvm-17/lib/clang/17*/lib/linux/liborc_rt*" (tried in ., debian/tmp)

i386 is not in the ALL_ORC_SUPPORTED_ARCH list. Previously, the package
shipped liborc_rt-x86_64.a, but since we don't ship amd64 multilib
binaries anymore, remove the reference.

Addresses regressions introduced with commit:
  400d89390891a1e95eb15837baca64bf000a352b
  Enable COMPILER_RT_DEFAULT_TARGET_ONLY for all archs